@@ -1,10 +1,10 @@
 {
   "name": "@ubermanu/audience",
   "private": true,
-  "version": "3.0.2",
+  "version": "3.1.0",
   "type": "module",
   "description": "Crowd reaction soundboard",
-  "author": "Emmanuel Vodor <manu@paca.happy-dev.fr>",
+  "author": "Emmanuel Vodor <emmanuel.vodor@proton.me>",
   "repository": {
     "type": "git",
     "url": "https://github.com/ubermanu/audience.git"
@@ -19,25 +19,23 @@
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./jsconfig.json --watch"
   },
   "devDependencies": {
-    "@sveltejs/adapter-auto": "^3.0.1",
-    "@sveltejs/adapter-static": "^3.0.1",
-    "@sveltejs/kit": "^2.0.6",
-    "@sveltejs/vite-plugin-svelte": "^3.0.1",
-    "@ubermanu/prettier-config": "^3.2.0",
-    "@vite-pwa/sveltekit": "^0.3.0",
+    "@sveltejs/adapter-static": "^3.0.6",
+    "@sveltejs/kit": "^2.9.0",
+    "@sveltejs/vite-plugin-svelte": "^5.0.1",
+    "@tailwindcss/vite": "4.0.0-beta.4",
+    "@ubermanu/prettier-config": "^3.4.0",
+    "@vite-pwa/sveltekit": "^0.6.6",
     "animate.css": "^4.1.1",
-    "autoprefixer": "^10.4.16",
-    "postcss": "^8.4.32",
-    "prettier": "^3.1.1",
-    "prettier-plugin-svelte": "^3.1.2",
-    "prettier-plugin-tailwindcss": "^0.5.9",
-    "svelte": "5.0.0-next.26",
-    "svelte-check": "^3.6.2",
-    "tailwindcss": "^3.4.0",
-    "typescript": "^5.3.3",
-    "vite": "^5.0.10",
-    "vite-plugin-pwa": "^0.17.4",
-    "workbox-window": "^7.0.0"
+    "prettier": "^3.4.1",
+    "prettier-plugin-svelte": "^3.3.2",
+    "prettier-plugin-tailwindcss": "^0.6.9",
+    "svelte": "5.5.0",
+    "svelte-check": "^4.1.0",
+    "tailwindcss": "4.0.0-beta.4",
+    "typescript": "^5.7.2",
+    "vite": "^6.0.2",
+    "vite-plugin-pwa": "^0.21.1",
+    "workbox-window": "^7.3.0"
   },
   "prettier": "@ubermanu/prettier-config"
 }
